项目上传数据处理ing

+++ module_common/src/main/java/com/sribs/common/bean/net/ProjectListRes.kt
@@ -14,8 +14,8 @@ data class ProjectListRes(
     @SerializedName("leaderName")        var leaderName:String,
     @SerializedName("buildingNo")        var buildingNo:String, //三期项目列表无楼号
     @SerializedName("inspectors")        var inspectors:List<String>?,
-    @SerializedName("parentVersion")     var parentVersion:Int,
-    @SerializedName("version")           var version:Int,
+    @SerializedName("parentVersion")     var parentVersion:Long,
+    @SerializedName("version")           var version:Long,
     @SerializedName("createTime")        var createTime:String,
     @SerializedName("updateTime")        var updateTime:String
 )
